+++ folenski/staff/Drivers/Sqlite.php
@@ -16,8 +16,9 @@ class Sqlite implements DriversInterface
     /**
      * @return string retourne le type json
      */
-    function typeJson(): string
+    function typeJson(bool $small = true): string
     {
+        if ($small) return "VARCHAR(" . self::SZ_SM_JSON . ")";
         return "VARCHAR(" . self::SZ_JSON . ")";
     }