+++ folenski/staff/Models/Message.php
@@ -6,6 +6,7 @@
  * @author  folenski
  * @version 1.0 09/08/2022 : Version Initiale 
  * @version 1.1 10/12/2022: utilisation des tags
+ * @version 1.2 09/07/2023: propriété _error ajoutée pour supprimer un warning
  * 
  */
 
@@ -21,12 +22,14 @@ final class Message implements TableInterface
         "user"       => "#TXT_SM NOT NULL",
         "host"       => "#TXT_SM",
         "hash"       => "#TXT_SM NOT NULL",
+        "spam"       => "INTEGER DEFAULT 0",
         "j_msg"      => "#JSON NOT NULL",
         "done"       => "INTEGER DEFAULT 0",
         "created_at" => "DATETIME NOT NULL",
         "_key"       => "FOREIGN KEY (user) REFERENCES %suser(user)",
         "_index"     => "hash"
     ];
+    private string|bool $_error = false;
 
     /**
      * @return array retourne le nom et la description de la table
@@ -52,7 +55,7 @@ final class Message implements TableInterface
      */
     function errors(): false|string
     {
-        return (isset($this->_error)) ? $this->_error : false;
+        return $this->_error;
     }
 
     /**